requirements.txt
SQLAlchemy==1.1.9
pycrypto==2.6.1
pymongo==3.6.0
Flask==0.12.2
flask_sqlalchemy==2.3.2
sqlalchemy_migrate==0.11.0
simplejson==3.13.2
